added rainbow mode
